fix: Display possession time with decimal precision
- Change possession parsing from parseInt() to parseFloat() to preserve decimals
- Add formatPossession() helper to show tenths when < 10s, whole seconds otherwise
- Update Main.vue and Score.vue to use formatPossession() for display
- Ensures accurate possession timer display in last 10 seconds (e.g., 9.5s, 3.2s)

Files modified:
- useWebSocket.js: parseFloat() for WebSocket message processing
- useGame.js: parseFloat() for HTTP fetch processing
- useFormat.js: new formatPossession() helper function
- Main.vue, Score.vue: use formatPossession() in templates

+++ sources/app2_live/composables/useGame.js
@@ -114,7 +114,7 @@ export const useGame = () => {
       // Update timer in store
       gameStore.updateScore({
         timer: timer,
-        possession: parseInt(data.possession) || 0,
+        possession: parseFloat(data.possession) || 0,
         period: data.period || gameStore.score.period
       })
     } catch (err) {
